changed code into input question and added difficulty level folder
